work loaded

+++ app/Services/RoleServices.php
@@ -2,81 +2,128 @@
 
 namespace App\Services;
 
-use App\Services\BaseServices;
+use App\Traits\Jsonify;
+use Exception;
 use Illuminate\Support\Facades\DB;
-use Illuminate\Support\Facades\Hash;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
 
 class RoleServices extends BaseServices
 {
+    use Jsonify;
+
     public function __construct(Role $model)
     {
         parent::__construct($model);
     }
+
     public function search()
     {
-        $model = $this->model;
-        return $this->model->paginate(10);
+        DB::beginTransaction();
+        try {
+            $model = $this->model;
+            $model = $this->model->paginate(10);
+            DB::commit();
+
+            return self::jsonSuccess(message: '', data: $model);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
+        }
     }
+
     public function getPermission()
     {
-        $permission = Permission::get();
-        return $permission;
+        DB::beginTransaction();
+        try {
+            $permission = Permission::get();
+            DB::commit();
+
+            return self::jsonSuccess(message: '', data: $permission);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
+        }
     }
+
     public function store($request)
     {
+        DB::beginTransaction();
         try {
             $role = Role::create(['name' => $request->input('name')]);
             $role->syncPermissions($request->input('permission'));
-            return $role;
-        } catch (\Throwable $th) {
-            throw $th;
+            DB::commit();
+
+            return self::jsonSuccess(message: '', data: $role);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
         }
     }
+
     public function show($role)
     {
+        DB::beginTransaction();
         try {
             $id = $role->id;
             $role = Role::find($id);
-            $rolePermissions = Permission::join("role_has_permissions", "role_has_permissions.permission_id", "=", "permissions.id")
-                ->where("role_has_permissions.role_id", $id)
+            $rolePermissions = Permission::join('role_has_permissions', 'role_has_permissions.permission_id', '=', 'permissions.id')
+                ->where('role_has_permissions.role_id', $id)
                 ->get();
-            return $rolePermissions;
-        } catch (\Throwable $th) {
-            throw $th;
+            DB::commit();
+
+            return self::jsonSuccess(message: '', data: $rolePermissions);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
         }
     }
+
     public function edit($role)
     {
+        DB::beginTransaction();
         try {
             $id = $role->id;
             $role = Role::find($id);
             $permission = Permission::get();
-            $rolePermissions = DB::table("role_has_permissions")->where("role_has_permissions.role_id", $id)
+            $rolePermissions = DB::table('role_has_permissions')->where('role_has_permissions.role_id', $id)
                 ->pluck('role_has_permissions.permission_id', 'role_has_permissions.permission_id')
                 ->all();
             $permission = [$role, $permission, $rolePermissions];
-            return $permission;
-        } catch (\Throwable $th) {
-            throw $th;
+            DB::commit();
+
+            return self::jsonSuccess(message: '', data: $rolePermissions);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
         }
     }
+
     public function update($role, $request)
     {
         try {
-
         } catch (\Throwable $th) {
             throw $th;
         }
     }
+
     public function deleteRole($role)
     {
+        DB::beginTransaction();
         try {
-            DB::table("roles")->where('id',$role->id)->delete();
-            return ;
-        } catch (\Throwable $th) {
-            //throw $th;
+            $rolePermissions = DB::table('roles')->where('id', $role->id)->delete();
+            DB::commit();
+
+            return self::jsonSuccess(message: 'Permissions deleted', data: $rolePermissions);
+        } catch (Exception $exception) {
+            DB::rollback();
+
+            return self::jsonError($exception->getMessage());
         }
     }
 }
